backend/requirements.txt
flask==2.0.1
pandas==1.3.0
scikit-learn==0.24.2
joblib==1.3.2
flask-cors==3.0.10
flask-sqlalchemy==3.0.2
flask-jwt-extended==4.4.4
werkzeug==2.0.1
python-dotenv==0.19.0
numpy==1.21.0